requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
psycopg2-binary==2.9.1
pytest==6.2.5
pydantic==1.8.2
